Number of render commands configurable
I realized that I can use template arguments to specify the statically
allocated number of render commands. This makes it simpler to decide the
number of render commands to be used in the project level and not globally.

Ironically I thus had to move the code into the header again.